Clean up searches for GLib submodules on Windows.
FindWSWinLibs() does nothing on non-Windows platforms; only use it on
Windows.

If we aren't building the GTK+ version of Windows, search for the GLib
modules in gtk3 first, and then gtk2; don't depend on which version of
GTK+ was enabled, as we presumably didn't enable *either* of them.

While we're at it, fix some apparent typoes in FindGTHREAD2.cmake - it
made references to GMODULE2_LIBRARY_DIRS that presumably should have
been references to GTHREAD2_LIBRARY_DIRS.

Change-Id: I418ac71803cfca2fd0ad99014d995241819ddc38
Reviewed-on: https://code.wireshark.org/review/32039
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>