Don't allow NSIS or WiX installers if we aren't building Wireshark.
In both cases, a file used in the process of building the installer is
generated from the Wireshark binary, so, unless we also arrange not to
put that file in the installer if Wireshark isn't being built, we can't
build the installer.

Have ADD_NSIS_PACKAGE_TARGET and ADD_WIX_PACKAGE_TARGET check whether
we're building Wireshark and fail if we aren't.  That way, *if* we're
including the NSIS or WiX packages in the build files, we fail if we
aren't building Wireshark.

Change-Id: Icfe4d2491bd721cdd5fef424e9a7565c4990defa
Reviewed-on: https://code.wireshark.org/review/32053
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>