Don't fail in CMake on Windows if you've disabled building Wireshark.
It's possible to do a *build* on Windows without the Wireshark app being
built; however, the *installers* can't currently be built.

If we have NSIS or WiX, and Wireshark isn't being built, report a
warning, and don't have targets for building the installers.

Change-Id: I2cc9c1f6ba375dbcb6d5b7520d2fa33ad97ba9fe
Reviewed-on: https://code.wireshark.org/review/32061
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>