Fix liblzma, liblz4, and libssh builds with minimum deployment version and SDK.
liblzma is used by libxml2, which is used by libwireshark, so we need
it.  It's build by xz, so we need to do the xz build with the
minimum-version and SDK flags.

liblz4 does its builds in its own unique way; it appears we need to set
MOREFLAGS to include the relevant flas.

libssh is built with CMake, so you have to do special magic to set the
SDK path; do it by running cmake with the MACOSX_DEPLOYMENT_TARGET and
SDKROOT environment variables.

Change-Id: Ia588632e5047c4a8a22fe6ef3a0844cfe7722df2
Reviewed-on: https://code.wireshark.org/review/32171
Reviewed-by: Guy Harris <guy@alum.mit.edu>